supply_chain_streamlit_project/requirements.txt
streamlit
pandas
plotly
faker
python-dateutil
